Code Update

+++ src/MicroService.Service/Services/CalculationService.cs
@@ -15,7 +15,7 @@ namespace MicroService.Service.Services
             _testDataRepository = testDataRepository ?? throw new ArgumentNullException(nameof(testDataRepository));
         }
 
-        public async Task<double> CalculatePercentile(double[] sequence, double excelPercentile)
+        public async Task<double> CalculatePercentile(float[] sequence, double excelPercentile)
         {
             var data = await _testDataRepository.FindAll();
             var array = data.Select(x => x.Data).ToArray();
